requirements.txt
fastapi
uvicorn
streamlit
pandas
scikit-learn
prophet
plotly
google-generativeai
python-dotenv
requests
SQLAlchemy
pydantic
joblib
uvicorn[standard]
sqlalchemy
alembic     # optional later if you want proper migrations
pytest
httpx
